pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "OpenHosta"
version = "4.0.0"
description = "A lightweight library integrating LLM natively into Python"
keywords = ["AI", "GPT", "Natural language", "Autommatic", "Easy"]
authors = [
  { name="Léandre Ramos"},
  { name="Merlin Devillard"},
  { name="William Jolivet"},
  { name="Emmanuel Batt"},
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8,<=3.14"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: Implementation :: CPython",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "Natural Language :: French",
    "Natural Language :: English",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries",
    "Typing :: Typed"
]
dependencies = [
    "requests>=2.32.3",
    "typing_extensions>=4.12.2",
    "jinja2>=3.1.5",
    "python-dotenv>=1.1.1"
]

[project.optional-dependencies]
dev = [ 
    "mypy>=1.13.0",
    "isort>=5.13.2",
    "autopep8>=2.3.1",
    "pylint>=3.3.1",
    "pyflakes>=3.2.0",
    "bandit>=1.7.10"
]
tests = [
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
    "pillow>=11.0.0",
    "pydantic>=2.8.2",
    "Flask>=3.0.3"
]

[project.urls]
Homepage = "https://github.com/hand-e-fr/OpenHosta"
Issues = "https://github.com/hand-e-fr/OpenHosta/issues"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"OpenHosta" = ["utils/*.json"]

[tool.bandit]
skips = ["B112"]
targets = ["src/OpenHosta"]
recursive = true

[tool.pylint]
disable = [
    "C0111",  
    "C0103",  
    "C0114",  
    "C0115",  
    "C0116",  
    "R0903",
    "C0415"
]
max-line-length = 120

[tool.mypy]
disable_error_code = ["attr-defined"]